clang recommended changes incorporated

+++ src/adj_list.cpp
@@ -1248,7 +1248,8 @@ OutCursor *AdjList::get_outnbd_iter()
     OutCursor *toReturn =
         new AdjOutCursor(get_new_out_adjlist_cursor(), session);
     toReturn->set_num_nodes(num_nodes);
-    toReturn->set_key_range({-1, static_cast<node_id_t>(num_nodes - 1)});
+    toReturn->set_key_range(
+        {-1, static_cast<node_id_t>((int64_t)num_nodes - 1)});
     return toReturn;
 }
 
@@ -1257,7 +1258,8 @@ InCursor *AdjList::get_innbd_iter()
     uint64_t num_nodes = this->get_num_nodes();
     InCursor *toReturn = new AdjInCursor(get_new_in_adjlist_cursor(), session);
     toReturn->set_num_nodes(num_nodes);
-    toReturn->set_key_range({-1, static_cast<node_id_t>(num_nodes - 1)});
+    toReturn->set_key_range(
+        {-1, static_cast<node_id_t>((int64_t)num_nodes - 1)});
     return toReturn;
 }
